# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Autenticación",
        "description": "Documentación para la API con autenticación mediante Sanctum y Swagger",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener todos los productos",
                "operationId": "1bfaa78d1c2c3848ab8165c5dadcad3e",
                "responses": {
                    "200": {
                        "description": "Lista de productos paginada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 20
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear nuevo producto",
                "operationId": "e8185e307706e2cd84bbcf5171e2f09d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener los detalles de un producto",
                "operationId": "7ab21ddb713501f3d4c2914892c02a21",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del producto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualizar un producto existente",
                "operationId": "6a7e11900844cd85cff66e662b82bd4c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar un producto",
                "operationId": "e80a6ca46716acee6b47ea1ac91663c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Producto eliminado correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products/{productId}/prices": {
            "get": {
                "tags": [
                    "Precios"
                ],
                "summary": "Obtener todos los precios de un producto",
                "operationId": "9795682918251db3525a68361862426f",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de precios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProductPrice"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Precios"
                ],
                "summary": "Registrar un nuevo precio para un producto",
                "operationId": "e607c4b0afa1b3f2dfbb98a8dda32834",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "price",
                                    "tax_cost",
                                    "currency_id"
                                ],
                                "properties": {
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150
                                    },
                                    "tax_cost": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 20
                                    },
                                    "currency_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Precio creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductPrice"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProductPrice": {
                "title": "Precio de Producto",
                "required": [
                    "price",
                    "tax_cost",
                    "currency"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 150
                    },
                    "tax_cost": {
                        "type": "number",
                        "format": "float",
                        "example": 20
                    },
                    "currency_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "currency": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "USD"
                            },
                            "symbol": {
                                "type": "string",
                                "example": "$"
                            },
                            "exchange_rate": {
                                "type": "number",
                                "format": "float",
                                "example": 1
                            }
                        },
                        "type": "object"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-04-05T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-04-05T12:30:00Z"
                    }
                },
                "type": "object"
            },
            "Product": {
                "title": "Producto",
                "required": [
                    "name",
                    "description",
                    "manufacturing_cost",
                    "prices"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Producto A"
                    },
                    "description": {
                        "type": "string",
                        "example": "Descripción del producto A"
                    },
                    "manufacturing_cost": {
                        "type": "number",
                        "format": "float",
                        "example": 150
                    },
                    "prices": {
                        "type": "array",
                        "items": {
                            "required": [
                                "price",
                                "tax_cost",
                                "currency_id"
                            ],
                            "properties": {
                                "price": {
                                    "type": "number",
                                    "format": "float",
                                    "example": 200
                                },
                                "tax_cost": {
                                    "type": "number",
                                    "format": "float",
                                    "example": 20
                                },
                                "currency_id": {
                                    "type": "integer",
                                    "example": 1
                                }
                            },
                            "type": "object"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}